refactor: reorganize file structure and rename components
