+++ src/screens/stamp/StampBoard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, Text, StyleSheet, Image } from 'react-native';
-import colors from '../colors';
+import colors from '../../design/colors';
 
 type Stamp = {
   id: number;
